requirements.txt
fastapi
uvicorn
pydantic
requests
langchain
langchain-community
langchain-ollama
python-dotenv
tenacity
python-dotenv
